% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burritos_cost.R
\name{burrito_cost}
\alias{burrito_cost}
\title{Title calculte Total Burrito Cost}
\usage{
burrito_cost(price, count)
}
\arguments{
\item{price}{price per burrito in dollars}

\item{count}{number of burritos in order}
}
\description{
Title calculte Total Burrito Cost
}
